pyproject.toml
[project]
name = "sdc-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sdc_common_module",
    "sdc_archiver",
    "sdc_viewer",
    "py7zr>=1.1.0",
    "pycryptodome>=3.23.0"
]

[dependency-groups]
dev = [
    "pytest>=9.0.2"
]

[tool.uv]
package = false

[tool.uv.sources]
sdc_common_module = {workspace=true}
sdc_archiver = {workspace=true}
sdc_viewer = {workspace=true}

[tool.uv.workspace]
members = ["packages/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.pyright]
extraPaths = ["src"]
typeCheckingMode = "off"
venvPath = "."
venv = ".venv"
